added credits file
was looking at an old github repo, and this was something I
used to add to the top of my programs. I stopped for years
because I thought it was pretentious for someone just starting
out as a programmer, but looking back I now wish I had done it more.

It certainly wouldn't have mattered much since the bulk of my work
has been on private projects, but this is my first real public
contribution of software that wasn't a PR for another project
(where adding my own CREDITS.txt would be frowned upon).

Anyway, just a bit of info about why it's been added.
I've now been a programmer for 5 years, and am at the
time of this writing a 4th year computer science major.

In summary, I feel that I've earned my right to be a little
pretentious. I think we should all feel okay to look at our work
proudly and say 'yeah, I did that'.